Log GivethDirectory error
For local deploys, it the user hasn't updated the giveth directory
address, getState will keep throwing errors resulting in a web3monitor
reset and an infinate loop with no logs. Set a timeout on the reset, and
log the error to console so the user has a better chance of finding it.