add milestone w/o canceling those inProgress
This is a quick hack to allow adding a new milestone w/o canceling those
in progress. The contract needs to be fixed to implement this correctly.

This commit adds all inProgress milestones to the proposal when
proposing new milestones. When the proposed milestones are accepted, the
in progress milestones will be canceled, but they will be re-created b/c
they were included in the proposal.

There are some issues such as if an inProgress milestone is acted upon
(canceled, marked complete) after the newMilestones have been proposed,
but before they have been accepted, the milestone will be duplicated and appear
as inProgress upon acception.